+++ src/app/schools-detail/schools-detail.component.ts
@@ -37,8 +37,11 @@ export class SchoolsDetailComponent implements OnInit {
     });
   }
   removeClass(id){
-    this.classService.removeClass(id).subscribe(response =>{
+    if(confirm('Bạn có chắc chắn muốn xóa trường này ?')){
+      this.classService.removeClass(id).subscribe(response =>{
       this.classs = this.classs.filter(obj => obj.id != response.id);
   })
+    }
+    
   }
 }
